% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{beta_function}
\alias{beta_function}
\title{The beta function used in he EM algorithm}
\usage{
beta_function(para_beta, w)
}
\arguments{
\item{para_beta}{Vector of length 3 containing the parameters in beta function}

\item{w}{Scalar. The wealth in the beta function}
}
\description{
It is bascally a third polynomial of the wealth.
}
\author{
Yangzhuoran Yang
}
