% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{c_function}
\alias{c_function}
\title{The comsumption function used in the EM algorithm}
\usage{
c_function(para_c, w)
}
\arguments{
\item{para_c}{Vector of length 2 containing the parameters in the comsumption function}

\item{w}{Scalar. The Wealth in the comsumption function.}
}
\description{
It is the wealth times the logit function of a first order polynomial of wealth
}
\author{
Yangzhuoran Yang
}
