% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{iterate_EM}
\alias{iterate_EM}
\title{Wrapper to iterate E and M steps.}
\usage{
iterate_EM()
}
\arguments{
\item{n_iter}{Integer. The number of iteration.}
}
\description{
The function reads the parameters and value from the environment in which it runs,
which is normally the global Environment.
}
\author{
Yangzhuoran Yang
}
