% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{lasso_weights}
\alias{lasso_weights}
\title{Lasso portfolio selection}
\usage{
lasso_weights(Rt, N = NCOL(Rt[[1]]), qp_lasso = TRUE,
  qp_weights = NULL)
}
\description{
Default using lasso after quadratic programming
}
\author{
Yangzhuoran Yang
}
