% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_optim.R
\name{optim_dynam}
\alias{optim_dynam}
\title{Dynamic optimization}
\usage{
optim_dynam(Rt, M = NROW(Rt[[1]]), Tn = length(Rt) - 1, weights,
  W_ini = 1000, utility = power_u, beta = 1/1.05,
  consu_function = inv.logit)
}
\arguments{
\item{Rt}{returns}

\item{weights}{weight matrix where each column contain weights for one period}

\item{W_ini}{initial wealth currently set to be the same 1000}

\item{utility}{utility function}

\item{beta}{discount factor. Currently set to be the same 1/1.05}

\item{consu_function}{consumption has been set to inverse logit}
}
\value{
list contain theta and value
}
\description{
Dynamic optimization
}
\author{
Yangzhuoran Yang
}
