% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{round_EM}
\alias{round_EM}
\title{The iteration fucntion used in the EM algorithm}
\usage{
round_EM(Rr, Rf, valuefunction = value_varmean, M = NROW(Rr[[1]]),
  Tn = length(Rr) - 1, ini_W = 1000, discount = 1/1.01,
  lambda = 1/2, para = NULL)
}
\arguments{
\item{Rr}{List of matrix. The rertuns of the assets.
Each list is for one time point and each column in the matrix is one assets.}

\item{Rf}{Scalar. The risk free rate. e.g. 1.01}

\item{valuefunction}{Function. The function that returns the value being optimised.
The first argument needs to be the parameter to optimise.}

\item{M}{Integer. The number of instances for one assets at a time}

\item{Tn}{Integer. The length of time horizon. Not including 0.}

\item{ini_W}{Scalar. The initial wealth.}

\item{discount}{Scalar. The rate of discount. e.g. 1/1.01}

\item{lambda}{The scalar that make the calculation easiler.}

\item{para}{Matrix. Initial value of all the paramters. Each column is a parameter and each row is a time point.}
}
\description{
This is only for one round of E step and M step
}
\seealso{
\code{iterate_EM}
}
\author{
Yangzhuoran Yang
}
