% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim_simple}
\alias{sim_simple}
\title{Simple simulation}
\usage{
sim_simple(R1 = NULL, mu = NULL, vol = NULL, Tn = 10, N = 5,
  M = 10000, defaultsetting = TRUE)
}
\arguments{
\item{R1}{Returns at time 0. Should be a M*N matrix where M and N are the number of entry and number of assets.}

\item{mu}{Vector with length N contains mean of the assets returns}

\item{vol}{Vector with length N contains volatilities of the assets returns}

\item{Tn}{Number of periods, excluding time 0.}

\item{N}{number of assets}

\item{M}{number of realization}
}
\value{
List of returns
}
\description{
Simulate returns using Choleski Decomposition where correlation depends on the order of assets
}
\author{
Yangzhuoran Yang
}
