% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_weights.R
\name{weights_qp}
\alias{weights_qp}
\alias{weights_lasso}
\title{Additional portfolio selection weights method: LASSO, qp}
\usage{
weights_qp(Rt, constr = 1)

weights_lasso(Rt, N = NCOL(Rt[[1]]), qp_lasso = TRUE,
  qp_weights = NULL)
}
\arguments{
\item{Rt}{Lists of return rt}

\item{constr}{If numeric, take as the imposed constraint.
If NULL, no constraint. Otherwise compute proper constraint using 10-fold cross-validation}

\item{qp_lasso}{if TRUE, taking no-short-sale portfolio from quadratic programming as y in lasso}

\item{qp_weights}{can supply weights of o-short-sale portfolio from quadratic programming}
}
\value{
List of weights for each period
}
\description{
Default using lasso after quadratic programming. As a approximation to constrained risk
minimization
}
\section{Functions}{
\itemize{
\item \code{weights_qp}: portfolio selection quadratic programming for multiperiod data set
}}

\seealso{
\code{\link{weights_lm}}
}
\author{
Yangzhuoran Yang
}
