% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_optim.R
\name{value_func}
\alias{value_func}
\title{Value Function}
\usage{
value_func(updatetheta = NULL, i = 1, theta, Rt, M = NULL,
  Tn = length(Rt) - 1, weights, W = NULL, ValueOnly = FALSE,
  utility = power_u, beta = 1/1.05, consu_function = boot::inv.logit)
}
\arguments{
\item{updatetheta}{updated theta}

\item{i}{starting period}

\item{theta}{matrix contain all theta}

\item{Rt}{returns}

\item{M}{number of realization}

\item{Tn}{Number of periods, excluding time 0.}

\item{weights}{weight matrix where each column contain weights for one period}

\item{W}{wealth}

\item{ValueOnly}{logical. If FALSE, returns both value and wealth}

\item{utility}{utility function}

\item{beta}{discount factor. Currently set to be the same 1/1.05}

\item{consu_function}{consumption has been set to inverse logit}
}
\value{
theta
}
\description{
Value Function
}
\author{
Yangzhuoran Yang
}
